classifier/requirements.txt
pdfplumber>=0.11
pymupdf>=1.24
pypdf>=5.0
pyyaml>=6.0
pytest>=8.0
rapidfuzz>=3.0
